fix: rename refillToken to refill_token and fix CI lint workflow
- Rename ai/refillToken.py to ai/refill_token.py (N999 fix)
- Update import in extensions/loops.py
- Redesign ruff_linter.yml: auto-fix first, then commit, then check remaining issues
