+++ .github/workflows/ruff_linter.yml
@@ -26,14 +26,17 @@ jobs:
       - name: Install Ruff
         run: pip install ruff
 
-      - name: Run Ruff Linter
-        run: ruff check . --fix
+      - name: Run Ruff Linter (auto-fix)
+        run: ruff check . --fix --exit-zero
 
       - name: Run Ruff Formatter
         run: ruff format .
 
-      - name: Commit changes
+      - name: Commit auto-fixes
         uses: stefanzweifel/git-auto-commit-action@v5
         with:
           commit_message: "style: autoformat and lint Python code with Ruff"
           branch: ${{ github.head_ref }}
+
+      - name: Check for remaining lint issues
+        run: ruff check .
